requirements.txt
mkdocs
mkdocs-terminal
PyYAML
